Se modifica la cantida de double por int

+++ src/main/java/com/warriors/easyStock/Producto/entities/Producto.java
@@ -52,7 +52,7 @@ public class Producto implements Serializable {
     @NotNull
     @JsonProperty("cantidad_stock")
     @Column(name = "cantidad_stock", nullable = false)
-    private Double cantidadStock;
+    private int cantidadStock;
     @JsonProperty("fecha_creacion")
     @Temporal(TemporalType.DATE)
     @Column(name = "fecha_creacion")
